backend/requirements.txt
PyYAML==6.0.1
colorama==0.4.6